Hotels__TET_bot/requirements.txt
arrow==1.2.2
async-timeout==4.0.2
Deprecated==1.2.13
distlib==0.3.4
idna==3.3
loguru==0.6.0
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
packaging==21.3
params==0.9.0
pipenv==2022.4.21
pyparsing==3.0.8
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot-calendar==1.0.5
redis==4.2.2
requests==2.28.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
utils==1.0.1
virtualenv==20.14.1
win32-setctime==1.1.0
